+++ Tests/Unit/Microsoft365DSC/Microsoft365DSC.AADNamedLocationPolicy.Tests.ps1
@@ -60,14 +60,14 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                     Credential  = $credsGlobalAdmin
                 }
 
-                Mock -CommandName Get-MgIdentityConditionalAccessNamedLocation -MockWith {
+                Mock -CommandName Get-MgBEtaIdentityConditionalAccessNamedLocation -MockWith {
                     return $null
                 }
             }
 
             It 'Should return values from the get method' {
                 (Get-TargetResource @testParams).Ensure | Should -Be 'Absent'
-                Should -Invoke -CommandName 'Get-MgIdentityConditionalAccessNamedLocation' -Exactly 1
+                Should -Invoke -CommandName 'Get-MgBetaIdentityConditionalAccessNamedLocation' -Exactly 1
             }
             It 'Should return false from the test method' {
                 Test-TargetResource @testParams | Should -Be $false
@@ -87,7 +87,7 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                     Credential  = $credsGlobalAdmin
                 }
 
-                Mock -CommandName Get-MgIdentityConditionalAccessNamedLocation -MockWith {
+                Mock -CommandName Get-MgBetaIdentityConditionalAccessNamedLocation -MockWith {
                     return @{
                         DisplayName          = 'Company Network'
                         Id                   = '046956df-2367-4dd4-b7fd-c6175ec11cd5'
@@ -104,7 +104,7 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
 
             It 'Should return values from the get method' {
                 (Get-TargetResource @testParams).Ensure | Should -Be 'Present'
-                Should -Invoke -CommandName 'Get-MgIdentityConditionalAccessNamedLocation' -Exactly 1
+                Should -Invoke -CommandName 'Get-MgBetaIdentityConditionalAccessNamedLocation' -Exactly 1
             }
 
             It 'Should return false from the test method' {
@@ -128,7 +128,7 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                     Credential  = $credsGlobalAdmin
                 }
 
-                Mock -CommandName Get-MgIdentityConditionalAccessNamedLocation -MockWith {
+                Mock -CommandName BEta -MockWith {
                     return @{
                         DisplayName          = 'Company Network'
                         Id                   = '046956df-2367-4dd4-b7fd-c6175ec11cd5'
@@ -145,7 +145,7 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
 
             It 'Should return Values from the get method' {
                 Get-TargetResource @testParams
-                Should -Invoke -CommandName 'Get-MgIdentityConditionalAccessNamedLocation' -Exactly 1
+                Should -Invoke -CommandName 'Get-MgBetaIdentityConditionalAccessNamedLocation' -Exactly 1
             }
 
             It 'Should return true from the test method' {
@@ -164,7 +164,7 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                     Credential  = $credsGlobalAdmin
                 }
 
-                Mock -CommandName Get-MgIdentityConditionalAccessNamedLocation -MockWith {
+                Mock -CommandName Get-MgBetaIdentityConditionalAccessNamedLocation -MockWith {
                     return @{
                         DisplayName          = 'Company Network'
                         Id                   = '046956df-2367-4dd4-b7fd-c6175ec11cd5'
@@ -179,7 +179,7 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
 
             It 'Should return values from the get method' {
                 Get-TargetResource @testParams
-                Should -Invoke -CommandName 'Get-MgIdentityConditionalAccessNamedLocation' -Exactly 1
+                Should -Invoke -CommandName 'Get-MgBetaIdentityConditionalAccessNamedLocation' -Exactly 1
             }
 
             It 'Should return false from the test method' {
@@ -199,7 +199,7 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                     Credential = $Credential
                 }
 
-                Mock -CommandName Get-MgIdentityConditionalAccessNamedLocation -MockWith {
+                Mock -CommandName Get-MgBetaIdentityConditionalAccessNamedLocation -MockWith {
                     return @{
                         DisplayName          = 'Company Network'
                         Id                   = '046956df-2367-4dd4-b7fd-c6175ec11cd5'